requirements.txt
streamlit
langchain_groq 
langchain_anthropic
langchain_community 
pypdf 
sentence_transformers
python-dotenv
chromadb
torch
transformers